chore(commercetools:api-client): renamed graphqlFactory.ts to createCommerceToolsConnection.ts
